backend/requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.1.0
openai>=1.12.0

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.25.0

# Google APIs
google-api-python-client>=2.100.0
google-search-results>=2.4.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0